requirements.txt
Django==4.2.5
bpython==0.24
Pillow==10.0.1
path.py==12.5.0
django-annoying
django-extensions
mysqlclient==2.2.0
django-sekizai==3.0.1
social-auth-app-django
django_facebook
sorl-thumbnail

git+https://github.com/IlianIliev/django-ad-code.git

python-dateutil==2.8.2
sentry-sdk==1.32.0
fabric==3.1.0
unipath==1.1
django-debug-toolbar==4.1.0
werkzeug==3.0.1
